requirements.txt
pyyaml==6.0
deepdiff==5.8.1